requirements.txt
asgiref==3.2.7
celery==5.0.1
Django==3.0.7
django-admin-ordering==0.14.0
djangorestframework==3.12.4
gunicorn==20.0.4
ipython==7.19.0
psycopg2-binary==2.8.5
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1

